Dev jaeke

+++ viajatcc-frontend/src/app/components/bus/fornecedores/fornecedores.component.ts
@@ -1,6 +1,10 @@
 import { Component } from '@angular/core';
 import { BusService, supplier } from '../../../services/bus.service';
 import { Router } from '@angular/router';
+import { ConfirmationDialogComponent } from '../../../shared/components/confirmation-dialog/confirmation-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { FormFornecedorComponent } from '../form-fornecedor/form-fornecedor.component';
 
 @Component({
   selector: 'app-fornecedores',
@@ -12,30 +16,42 @@ export class FornecedoresComponent {
   suppliers: supplier[] = []
   isLoading = true;
   error = '';
-  constructor(private busService: BusService, private router: Router ) {
+  constructor(private busService: BusService, private router: Router, private dialog: MatDialog, private snackBar: MatSnackBar) {
     this.busService.getSuppliers().subscribe({
       next: (response) => {
         this.suppliers = response
         this.isLoading = false;
-    },
+      },
       error: (err) => {
         this.error = 'Erro ao carregar os fornecedores';
         this.isLoading = false;
         console.error(err);
       }
-  })
+    })
+  }
+  async createSupplier() {
+    this.dialog.open(FormFornecedorComponent, {width: '600px'}).afterClosed().toPromise()
+  }
+  updateSupplier(supplier:supplier) {
+    this.dialog.open(FormFornecedorComponent, {data: supplier, width: '600px'})
+  }
+  async deleteSupplier(id: string) {
+    const confirm = await this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja excluir este fornecedor?'
+    }).afterClosed().toPromise()
+    if (confirm) {
+      this.busService.deleteSupplier(id).subscribe({
+        next: (response) => {
+          window.location.reload()
+          this.snackBar.open('Fornecedor excluido', 'fechar', {
+            duration: 3000
+          })
+        }
+      })
+    }
   }
   viewSupplier(id: string) {
-    console.log(id);
-    
-  }
-  deleteSupplier(id: string) {
-    this.busService.deleteSupplier(id).subscribe({
-      next: (response)=> {
-        this.router.navigate(['/home/bus/list'])
-        console.log('excluido');
-        
-      } 
-    })
+    this.router.navigate([`/home/bus/${id}`])
+
   }
 }
